Migrate gRPC same server TLS test to new API
(cherry picked from commit bcc3d14039c966397e15e0b2a9ed1f574354a1a1)
