+++ http/grpc/src/main/resources/application.properties
@@ -9,13 +9,13 @@ quarkus.grpc.server.enable-reflection-service=true
 quarkus.grpc.clients.reflection-service.port=${quarkus.grpc.clients.plain.port}
 quarkus.grpc.clients.streaming.port=${quarkus.grpc.clients.plain.port}
 
-%ssl.quarkus.grpc.clients.plain.ssl.trust-store=tls/ca.pem
-%ssl.quarkus.grpc.clients.reflection-service.ssl.trust-store=${quarkus.grpc.clients.plain.ssl.trust-store}
-%ssl.quarkus.grpc.clients.streaming.ssl.trust-store=${quarkus.grpc.clients.plain.ssl.trust-store}
+%ssl.quarkus.grpc.clients.plain.ssl.trust-store=${grpc.client.ca-cert}
+%ssl.quarkus.grpc.clients.reflection-service.ssl.trust-store=${grpc.client.ca-cert}
+%ssl.quarkus.grpc.clients.streaming.ssl.trust-store=${grpc.client.ca-cert}
 
 # See https://github.com/quarkusio/quarkus/issues/38965 to learn, why we use these parameters
 %ssl.quarkus.grpc.clients.plain.port=${quarkus.http.ssl-port}
-%ssl.quarkus.http.ssl.certificate.files=tls/server.pem
-%ssl.quarkus.http.ssl.certificate.key-files=tls/server.key
-%ssl.quarkus.grpc.server.ssl.certificate=tls/server.pem
-%ssl.quarkus.grpc.server.ssl.key=tls/server.key
+%ssl.quarkus.http.ssl.certificate.files=${grpc.server.cert}
+%ssl.quarkus.http.ssl.certificate.key-files=${grpc.server.key}
+%ssl.quarkus.grpc.server.ssl.certificate=${grpc.server.cert}
+%ssl.quarkus.grpc.server.ssl.key=${grpc.server.key}
