--- http/grpc/src/main/resources/tls/ca.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDbTCCAlWgAwIBAgIUTbmeWmNplPK5ZjMS7M+eSFtN2YkwDQYJKoZIhvcNAQEL
-BQAwUDELMAkGA1UEBhMCQ1oxDDAKBgNVBAgMA0pNSzENMAsGA1UEBwwEQnJubzEP
-MA0GA1UECgwGUmVkSGF0MRMwEQYDVQQLDApRdWFya3VzLVFFMB4XDTI0MDIyNzA4
-MzA1MloXDTM0MDIyNDA4MzA1MlowUDELMAkGA1UEBhMCQ1oxDDAKBgNVBAgMA0pN
-SzENMAsGA1UEBwwEQnJubzEPMA0GA1UECgwGUmVkSGF0MRMwEQYDVQQLDApRdWFy
-a3VzLVFFMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAs/0x9/YIoueQ
-ckoItGRuW8CUc9kpdS35wfxcqrlyfQbo2i4idg5V9d4EYU3ONpvVl69O97t3DnXp
-XC90Mt+ActinPXs//ulxVelP5uVWo4nQILk889GHRfoIvsEKe+YZKTBxT92PJEDh
-qs1eDK5OJApX6ZRhHqZRjxVPRgwOhUE47qIHflDE0wX54zDPHGtwdSJtlDWSfbRg
-BLLzni50XAAPoEmHRb557yjYR6A3SfAE/Oaz++kh5HyDTrufGRQpn8R6MR2X8lJG
-FDSMqoCsf9AOtGzrpq8EoJb5hwBC7ko599beXb1A3PuYqQ4XU9gGCtTs3HHx2Am4
-t9fwS5OCmwIDAQABoz8wPTAMBgNVHRMEBTADAQH/MA4GA1UdDwEB/wQEAwICBDAd
-BgNVHQ4EFgQUo9s3q9xO/KVtZ4hiDyCsHmXWI+kwDQYJKoZIhvcNAQELBQADggEB
-AK/G72vu1xyITXJZeoi7pyj0iDk4nn0T6Znl5BaMZ5ukjxVH7EuIAOuCOOeDiYhe
-3WdwWHbxGdaZhJH9SF4CoCKNapVL1567lT0MU9JxfNWot8ipZ1J5WCHpor5HNl1D
-c072uEHQ7lTlErrvMDppPt6xb7dhXSVHneXve+RbyF8spzeKG31yg5DPkSiaIP8p
-qNnGw++J1Il0CQA11hYlH4wUE80atWEugTx04BHiK/HtYLQsbHKVSNw4gj7eFr//
-sQW0fliWPjmXLxAUU2efU/w6vs37LCuHApaN80yxTbK3J5LiPD8kH/OWRkPEcts9
-f2lZYZzQcMQvjbaSJ/cSzM0=
------END CERTIFICATE-----
