--- http/grpc/src/main/resources/tls/server.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCPCYGGsbW0zG6I
-TMvK85BF0jnd0L1U/XovBRCi3+VmrR2JovuazuKpmBwNyThsNLsj2KBLhYL6uwuB
-Kw0BeBW+yOaNB8gE7ef6MrT5qlp53X4XudmsucvdJnCSB01GC3GjElYe9rXrSuRH
-yPzU8QgpBAIJwFpbij/TWFUdynHUZ8oAjatzFlDzW4CoCTgo4NTdXhg5hsMSZvB0
-bkQ/jQF0PUvGCgbVthQGvSt75A69+ns8j2kxfb8DBrm94WpcxY1hMsVXP6/KZOyc
-rHlzVELAXJA0p0mHvokfQuU3xJT/Y5NbLDkq+9sh9sYjT5lw3tf0v99CrEOQgTC6
-jejADT2zAgMBAAECggEACKelGD7ZhVKXX5TyAdpCAQ+K49KVGjbqNT0juA8/JLrV
-3jWn7sKU8XkcPXNPADEin7UkYd12wvAdbpfpxgx7mFs0pBTz8+RnVHrL+41kwxn1
-Xr8mni5x2PRR/GwHr3TSz/C0mFQKRu31qShOsB3ThhPRgcCLVx2i5gliwRY2VAlK
-hZVmbhyC+4qF/BFNpzIIjyqNZsu8yjBP920DzX9TrQ7q0yQ5oE+Pyg5TMJPxUOiN
-dNiqNwVnQ77mz54eKJBDJbx431LS4Anp585AAruMj+7imGc7zbc4ZB4CtZ7x6f45
-3eBbiBhMepZ9wKg/Ym9aBdxxpy3gapZRs40DAJFC3QKBgQDBd1yINEnJ1ZZel/iA
-3rp8j/7rITz/bYIFkRqxihjRMSrdBKzTx3n3dnRKYjOh4B6CkWToaqRP3HPBVhKd
-pHjdz7Oo3PDrXnZcx41FUzYB7B0qkBD6vyqM9/U88da4ERtqNOczWCSHz/AQXfhc
-X5OlYA1Ed4tdLcyVqQL1/0Y7HwKBgQC9RVDFBh8vzMA5nr/gaZOwDee935Dmei7s
-AJqq6rRxkZEqSf7M090DUHaBTeg1jRMtGGNn6Q6LxaGe/jrRUvNTWVtOt1VfG8xZ
-r/3fM71VnhZ/m8D/rCOmNOYekKTITxNlBPGfDV9wUFFATNhyQCWQTbg4XmFzWQG7
-hEdalmk+7QKBgCMHU4+tt/Z9X5588ZeTvDw1bjhwajTtRO9xGF4w3NFzj4k5AXnO
-0jyGDAQzx5l1lNCbNqQGOv3ismq9BN3aG7A9nQ/kARL8pX2i++cja9HpSFaegxSD
-bFbdxl9kgjYNkuMl9P6M5QBaG+M6wG8pNvhobb6JzofudO5cDZcwwyyNAoGAWdtw
-rzlq0Py6PiDaI6a8ERdo8EIVvvY/FJhs1bw8ErbzXkpnB8OF6C7pNBZSqinh8sTj
-XM/OshkP1DYKopppHycLLGHpzA+cgvAE7VTZDK7TK548kKWe/yeaIOS29spkAM/K
-DqMArofTK13QXN2Ld+kODuTwCx00r1vrrFxAdzkCgYACF/hzTzT0BozFDT5Ve4sX
-fQBa4oZEiFOVqJ40ok3ICdqe22/38y5bQEoZ19TNmrK4nz/W+hOCpID26x/fel14
-MegjDjBfH77uwFgoI9O0uJiX4U89ZCOWmMXNH5wqqySJG+C+kF5ACONmGfaYrkPc
-8sjG1e5JXF2K5PxXHlRiIw==
------END PRIVATE KEY-----
