--- http/grpc/src/main/resources/tls/server.pem
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDOzCCAiMCFH7N97KjoaAbXGiiEZBxz9YkgRvZMA0GCSqGSIb3DQEBCwUAMFAx
-CzAJBgNVBAYTAkNaMQwwCgYDVQQIDANKTUsxDTALBgNVBAcMBEJybm8xDzANBgNV
-BAoMBlJlZEhhdDETMBEGA1UECwwKUXVhcmt1cy1RRTAeFw0yNDAyMjcwODMxMjda
-Fw0zNDAyMjQwODMxMjdaMGQxCzAJBgNVBAYTAkNaMQwwCgYDVQQIDANKTUsxDTAL
-BgNVBAcMBEJybm8xDzANBgNVBAoMBlJlZEhhdDETMBEGA1UECwwKUXVhcmt1cy1R
-RTESMBAGA1UEAwwJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEAjwmBhrG1tMxuiEzLyvOQRdI53dC9VP16LwUQot/lZq0diaL7ms7iqZgc
-Dck4bDS7I9igS4WC+rsLgSsNAXgVvsjmjQfIBO3n+jK0+apaed1+F7nZrLnL3SZw
-kgdNRgtxoxJWHva160rkR8j81PEIKQQCCcBaW4o/01hVHcpx1GfKAI2rcxZQ81uA
-qAk4KODU3V4YOYbDEmbwdG5EP40BdD1LxgoG1bYUBr0re+QOvfp7PI9pMX2/Awa5
-veFqXMWNYTLFVz+vymTsnKx5c1RCwFyQNKdJh76JH0LlN8SU/2OTWyw5KvvbIfbG
-I0+ZcN7X9L/fQqxDkIEwuo3owA09swIDAQABMA0GCSqGSIb3DQEBCwUAA4IBAQCM
-Yl1CFnKqP2LF/zZvdwiLkdaTbvefh4W0C31tLd2OaIGDo18cCr0OWAia2XD9f9f7
-dlTmDJhRT230S96/aR7FzT30OoyGFeuNq+C4M5d7lcwllKlG5zXupLl7D3l30fnf
-tUxUrFWbyh/xVGKRm4J2xP5MtIGOTfXBZsxqaawEN7U2bTmsA+/1vBWXJ+W2yCfs
-IoQPNH125wsDOiCvXDacn2jd+GxxZXtfv4UoZ3LZLGko5Tv4dubu1SwaC6oek2bc
-5trifNC9timIoKM0mqc4hdClB6YDDQ+pRLmy545B/EwP3xMmugnTkRo12miPLtGX
-TayJV9LQkMXHgn5tKTOp
------END CERTIFICATE-----
